requirements.txt
fastapi
langchain-ollama
pydantic
uvicorn